pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "econosim"
version = "0.1.0"
description = "Multi-agent AI economic simulation platform with stock-flow-consistent accounting"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "EconoSim Team"}]

dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "scipy>=1.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]
rl = [
    "gymnasium>=0.29",
    "stable-baselines3>=2.2",
    "torch>=2.1",
]
viz = [
    "streamlit>=1.29",
    "plotly>=5.18",
    "matplotlib>=3.8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.setuptools.packages.find]
where = ["."]
include = ["econosim*"]
